move most code to networkconfig module, simplify main
also: only initialize 2 WLAN instances (one AP, one STA)
